Added proper logout and auth to client endpoints

+++ backend/src/main/clients/DeleteClientService.js
@@ -13,6 +13,12 @@ class DeleteClientService {
   }
 
   deleteClient(id, loginData) {
+    const isAdmin = this._hasRole('admin', loginData);
+    const isFacilitator = this._hasRole('facilitator', loginData);
+    if (!isAdmin && !isFacilitator) {
+      console.log('WARNING: unauthorised attempt to create client', loginData);
+      return Promise.reject(new Error('Not authorised'));
+    }
     const connection = this._databaseManager.createConnection();
     let deletePromise = this._deleteClient(id, loginData, connection);
     const closeConnection = () =>
@@ -22,6 +28,10 @@ class DeleteClientService {
 
   _deleteClient(id, loginData, connection) {
     return this._clientRepository.delete(id, connection);
+  }
+
+  _hasRole(role, loginData) {
+    return loginData && loginData.roles.indexOf(role) >= 0;
   }
 }
